refactor(RacingCarGame): 기능 캡슐화
- RacingCarGame의 메서드 일부를 Cars로 이동
